% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riesz_lr_finder.R
\name{riesz_lr_finder}
\alias{riesz_lr_finder}
\title{Title}
\usage{
riesz_lr_finder(
  x,
  new_x,
  net = NULL,
  weights,
  .f,
  end_lr = 0.1,
  batch_size = 32,
  shuffle = TRUE,
  weight_decay = 0,
  optimizer = torch::optim_adam
)
}
\arguments{
\item{optimizer}{}
}
\description{
Title
}
