% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riesznet.R
\name{riesznet}
\alias{riesznet}
\title{Title}
\usage{
riesznet(
  data,
  shifted,
  .f,
  net = NULL,
  weights = NULL,
  max_lr = 0.1,
  epochs = 500,
  batch_size = 32,
  valid_ratio = 0.2,
  shuffle = TRUE,
  patience = 10,
  weight_decay = 0,
  optimizer = torch::optim_adam,
  verbose = TRUE
)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
